requirements.txt
web3
py-solc-x
python-dotenv